+++ CHANGES.md
@@ -1,12 +1,22 @@
-flycheck-perl6 Changes
+flycheck-raku Changes
 ===================
 
-0.1 (2015-04-??)
+0.3 (2021-03-07)
 ----------------
 
-- First release
+- Add project root lib to the include path
+- Error pattern improvements:
+  - Show full error messages (some error sentences were partially matched)
+  - Show multi-line error messages
+  - Show potential difficulties
 
 0.2 (2020-04-22)
 ----------------
 
 - Rename Perl 6 to Raku
+
+
+0.1 (2015-04-??)
+----------------
+
+- First release